pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "photome"
version = "0.1.0"
description = "Integrated local-first personal photo server with semantic search groundwork"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
  "fastapi>=0.115,<1.0",
  "numpy>=2.0,<3.0",
  "opencv-python-headless>=4.10,<5.0",
  "pillow>=11.0,<12.0",
  "pytesseract>=0.3.13,<1.0",
  "sqlalchemy>=2.0,<3.0",
  "uvicorn[standard]>=0.30,<1.0",
]

[project.optional-dependencies]
clip = [
  "open_clip_torch>=2.29,<3.0",
  "torch>=2.5,<3.0",
  "torchvision>=0.20,<1.0",
]
test = [
  "httpx>=0.27,<1.0",
  "pytest>=8.0,<9.0",
]

[project.scripts]
photome = "app.main:main"

[tool.setuptools]
include-package-data = false

[tool.setuptools.packages.find]
include = ["app*"]
